kvstore: Split logic into pkg/allocator
This is an intermediate commit. Some logic from pkg/kvstore/allocator has
been moved into pkg/allocator. No functionality should change, and most
changes are there to correct for package changes.
Unfortunately, the test logic is difficult to refactor this way. I have
commented it out since it is fixed in the next commit.
