+++ pkg/identity/cache/cache.go
@@ -19,10 +19,11 @@ import (
 	"reflect"
 
 	"github.com/cilium/cilium/api/v1/models"
+	"github.com/cilium/cilium/pkg/allocator"
 	"github.com/cilium/cilium/pkg/identity"
 	"github.com/cilium/cilium/pkg/idpool"
 	"github.com/cilium/cilium/pkg/kvstore"
-	"github.com/cilium/cilium/pkg/kvstore/allocator"
+	kvstoreallocator "github.com/cilium/cilium/pkg/kvstore/allocator"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/logging"
 	"github.com/cilium/cilium/pkg/logging/logfields"
@@ -51,7 +52,7 @@ func GetIdentityCache() IdentityCache {
 
 	if IdentityAllocator != nil {
 
-		IdentityAllocator.ForeachCache(func(id idpool.ID, val allocator.AllocatorKey) {
+		IdentityAllocator.ForeachCache(func(id idpool.ID, val kvstoreallocator.AllocatorKey) {
 			if val != nil {
 				if gi, ok := val.(globalIdentity); ok {
 					cache[identity.NumericIdentity(id)] = gi.LabelArray
@@ -80,7 +81,7 @@ func GetIdentityCache() IdentityCache {
 func GetIdentities() IdentitiesModel {
 	identities := IdentitiesModel{}
 
-	IdentityAllocator.ForeachCache(func(id idpool.ID, val allocator.AllocatorKey) {
+	IdentityAllocator.ForeachCache(func(id idpool.ID, val kvstoreallocator.AllocatorKey) {
 		if gi, ok := val.(globalIdentity); ok {
 			identity := identity.NewIdentityFromLabelArray(identity.NumericIdentity(id), gi.LabelArray)
 			identities = append(identities, identity.GetModel())
