+++ pkg/kvstore/allocator/allocator_test.go
@@ -16,526 +16,411 @@
 
 package allocator
 
-import (
-	"context"
-	"fmt"
-	"path"
-	"testing"
-	"time"
-
-	"github.com/cilium/cilium/pkg/idpool"
-	"github.com/cilium/cilium/pkg/kvstore"
-	"github.com/cilium/cilium/pkg/testutils"
-
-	. "gopkg.in/check.v1"
-)
-
-const (
-	testPrefix = "test-prefix"
-)
-
-func Test(t *testing.T) {
-	TestingT(t)
-}
-
+//import (
+//	"context"
+//	"fmt"
+//	"path"
+//	"testing"
+//	"time"
+//
+//	"github.com/cilium/cilium/pkg/allocator"
+//	"github.com/cilium/cilium/pkg/idpool"
+//	"github.com/cilium/cilium/pkg/kvstore"
+//	"github.com/cilium/cilium/pkg/testutils"
+//
+//	. "gopkg.in/check.v1"
+//)
+//
+//const (
+//	testPrefix = "test-prefix"
+//)
+//
+//func Test(t *testing.T) {
+//	TestingT(t)
+//}
+//
 type AllocatorSuite struct{}
 
-type AllocatorEtcdSuite struct {
-	AllocatorSuite
-}
-
-var _ = Suite(&AllocatorEtcdSuite{})
-
-func (e *AllocatorEtcdSuite) SetUpTest(c *C) {
-	kvstore.SetupDummy("etcd")
-}
-
-func (e *AllocatorEtcdSuite) TearDownTest(c *C) {
-	kvstore.DeletePrefix(testPrefix)
-	kvstore.Close()
-}
-
-type AllocatorConsulSuite struct {
-	AllocatorSuite
-}
-
-var _ = Suite(&AllocatorConsulSuite{})
-
-func (e *AllocatorConsulSuite) SetUpTest(c *C) {
-	kvstore.SetupDummy("consul")
-}
-
-func (e *AllocatorConsulSuite) TearDownTest(c *C) {
-	kvstore.DeletePrefix(testPrefix)
-	kvstore.Close()
-}
-
-type TestType string
-
-func (t TestType) GetKey() string { return string(t) }
-func (t TestType) String() string { return string(t) }
-func (t TestType) PutKey(v string) (AllocatorKey, error) {
-	return TestType(v), nil
-}
-
-func randomTestName() string {
-	return testutils.RandomRuneWithPrefix(testPrefix, 12)
-}
-
-func (s *AllocatorSuite) TestSelectID(c *C) {
-	allocatorName := randomTestName()
-	minID, maxID := idpool.ID(1), idpool.ID(5)
-	a, err := NewAllocator(allocatorName, TestType(""), WithMin(minID), WithMax(maxID), WithSuffix("a"))
-	c.Assert(err, IsNil)
-	c.Assert(a, Not(IsNil))
-
-	// allocate all available IDs
-	for i := minID; i <= maxID; i++ {
-		id, val, unmaskedID := a.selectAvailableID()
-		c.Assert(id, Not(Equals), idpool.NoID)
-		c.Assert(val, Equals, id.String())
-		c.Assert(id, Equals, unmaskedID)
-		a.mainCache.cache[id] = TestType(fmt.Sprintf("key-%d", i))
-	}
-
-	// we should be out of IDs
-	id, val, unmaskedID := a.selectAvailableID()
-	c.Assert(id, Equals, idpool.ID(0))
-	c.Assert(id, Equals, unmaskedID)
-	c.Assert(val, Equals, "")
-}
-
-func (s *AllocatorSuite) TestPrefixMask(c *C) {
-	allocatorName := randomTestName()
-	minID, maxID := idpool.ID(1), idpool.ID(5)
-	a, err := NewAllocator(allocatorName, TestType(""), WithMin(minID),
-		WithMax(maxID), WithSuffix("a"), WithPrefixMask(1<<16))
-	c.Assert(err, IsNil)
-	c.Assert(a, Not(IsNil))
-
-	// allocate all available IDs
-	for i := minID; i <= maxID; i++ {
-		id, val, unmaskedID := a.selectAvailableID()
-		c.Assert(id, Not(Equals), idpool.NoID)
-		c.Assert(id>>16, Equals, idpool.ID(1))
-		c.Assert(id, Not(Equals), unmaskedID)
-		c.Assert(val, Equals, id.String())
-	}
-
-	a.Delete()
-}
-
-func (s *AllocatorSuite) BenchmarkAllocate(c *C) {
-	allocatorName := randomTestName()
-	maxID := idpool.ID(256 + c.N)
-	allocator, err := NewAllocator(allocatorName, TestType(""), WithMax(maxID), WithSuffix("a"))
-	c.Assert(err, IsNil)
-	c.Assert(allocator, Not(IsNil))
-	defer allocator.DeleteAllKeys()
-
-	c.ResetTimer()
-	for i := 0; i < c.N; i++ {
-		_, _, err := allocator.Allocate(context.Background(), TestType(fmt.Sprintf("key%04d", i)))
-		c.Assert(err, IsNil)
-	}
-	c.StopTimer()
-
-}
-
-func (s *AllocatorSuite) TestGC(c *C) {
-	allocatorName := randomTestName()
-	maxID := idpool.ID(256 + c.N)
-	allocator, err := NewAllocator(allocatorName, TestType(""), WithMax(maxID), WithSuffix("a"), WithoutGC())
-	c.Assert(err, IsNil)
-	c.Assert(allocator, Not(IsNil))
-	defer allocator.DeleteAllKeys()
-	defer allocator.Delete()
-
-	allocator.DeleteAllKeys()
-
-	shortKey := TestType("1;")
-	shortID, _, err := allocator.Allocate(context.Background(), shortKey)
-	c.Assert(err, IsNil)
-	c.Assert(shortID, Not(Equals), 0)
-
-	longKey := TestType("1;2;")
-	longID, _, err := allocator.Allocate(context.Background(), longKey)
-	c.Assert(err, IsNil)
-	c.Assert(longID, Not(Equals), 0)
-
-	allocator.Release(context.Background(), shortKey)
-
-	keysToDelete := map[string]uint64{}
-	keysToDelete, err = allocator.RunGC(keysToDelete)
-	c.Assert(err, IsNil)
-	c.Assert(len(keysToDelete), Equals, 1)
-	keysToDelete, err = allocator.RunGC(keysToDelete)
-	c.Assert(err, IsNil)
-	c.Assert(len(keysToDelete), Equals, 0)
-
-	// wait for cache to be updated via delete notification
-	c.Assert(testutils.WaitUntil(func() bool { return allocator.mainCache.getByID(shortID) == nil }, 5*time.Second), IsNil)
-
-	key, err := allocator.GetByID(shortID)
-	c.Assert(err, IsNil)
-	c.Assert(key, Equals, TestType(""))
-}
-
-func testAllocator(c *C, maxID idpool.ID, allocatorName string, suffix string) {
-	allocator, err := NewAllocator(allocatorName, TestType(""), WithMax(maxID),
-		WithSuffix(suffix), WithoutGC())
-	c.Assert(err, IsNil)
-	c.Assert(allocator, Not(IsNil))
-
-	// remove any keys which might be leftover
-	allocator.DeleteAllKeys()
-
-	// allocate all available IDs
-	for i := idpool.ID(1); i <= maxID; i++ {
-		key := TestType(fmt.Sprintf("key%04d", i))
-		id, new, err := allocator.Allocate(context.Background(), key)
-		c.Assert(err, IsNil)
-		c.Assert(id, Not(Equals), 0)
-		c.Assert(new, Equals, true)
-
-		// refcnt must be 1
-		c.Assert(allocator.localKeys.keys[key.GetKey()].refcnt, Equals, uint64(1))
-	}
-
-	saved := allocator.backoffTemplate.Factor
-	allocator.backoffTemplate.Factor = 1.0
-
-	// we should be out of id space here
-	_, new, err := allocator.Allocate(context.Background(), TestType(fmt.Sprintf("key%04d", maxID+1)))
-	c.Assert(err, Not(IsNil))
-	c.Assert(new, Equals, false)
-
-	allocator.backoffTemplate.Factor = saved
-
-	// allocate all IDs again using the same set of keys, refcnt should go to 2
-	for i := idpool.ID(1); i <= maxID; i++ {
-		key := TestType(fmt.Sprintf("key%04d", i))
-		id, new, err := allocator.Allocate(context.Background(), key)
-		c.Assert(err, IsNil)
-		c.Assert(id, Not(Equals), 0)
-		c.Assert(new, Equals, false)
-
-		// refcnt must now be 2
-		c.Assert(allocator.localKeys.keys[key.GetKey()].refcnt, Equals, uint64(2))
-	}
-
-	// Create a 2nd allocator, refill it
-	allocator2, err := NewAllocator(allocatorName, TestType(""), WithMax(maxID),
-		WithSuffix("b"), WithoutGC())
-	c.Assert(err, IsNil)
-	c.Assert(allocator2, Not(IsNil))
-
-	// allocate all IDs again using the same set of keys, refcnt should go to 2
-	for i := idpool.ID(1); i <= maxID; i++ {
-		key := TestType(fmt.Sprintf("key%04d", i))
-		id, new, err := allocator2.Allocate(context.Background(), key)
-		c.Assert(err, IsNil)
-		c.Assert(id, Not(Equals), 0)
-		c.Assert(new, Equals, false)
-
-		localKey := allocator2.localKeys.keys[key.GetKey()]
-		c.Assert(localKey, Not(IsNil))
-
-		// refcnt in the 2nd allocator is 1
-		c.Assert(localKey.refcnt, Equals, uint64(1))
-
-		allocator2.Release(context.Background(), key)
-	}
-
-	// release 2nd reference of all IDs
-	for i := idpool.ID(1); i <= maxID; i++ {
-		allocator.Release(context.Background(), TestType(fmt.Sprintf("key%04d", i)))
-	}
-
-	// refcnt should be back to 1
-	for i := idpool.ID(1); i <= maxID; i++ {
-		key := TestType(fmt.Sprintf("key%04d", i))
-		c.Assert(allocator.localKeys.keys[key.GetKey()].refcnt, Equals, uint64(1))
-	}
-
-	keysToDelete := map[string]uint64{}
-	// running the GC should not evict any entries
-	keysToDelete, err = allocator.RunGC(keysToDelete)
-	c.Assert(err, IsNil)
-	c.Assert(len(keysToDelete), Equals, 0)
-
-	v, err := kvstore.ListPrefix(allocator.idPrefix)
-	c.Assert(err, IsNil)
-	c.Assert(len(v), Equals, int(maxID))
-
-	// release final reference of all IDs
-	for i := idpool.ID(1); i <= maxID; i++ {
-		allocator.Release(context.Background(), TestType(fmt.Sprintf("key%04d", i)))
-	}
-
-	for i := idpool.ID(1); i <= maxID; i++ {
-		key := TestType(fmt.Sprintf("key%04d", i))
-		c.Assert(allocator.localKeys.keys[key.GetKey()], IsNil)
-	}
-
-	// running the GC should evict all entries
-	keysToDelete, err = allocator.RunGC(keysToDelete)
-	c.Assert(err, IsNil)
-	c.Assert(len(keysToDelete), Equals, int(maxID))
-	keysToDelete, err = allocator.RunGC(keysToDelete)
-	c.Assert(err, IsNil)
-	c.Assert(len(keysToDelete), Equals, 0)
-
-	v, err = kvstore.ListPrefix(allocator.idPrefix)
-	c.Assert(err, IsNil)
-	c.Assert(len(v), Equals, 0)
-
-	allocator.DeleteAllKeys()
-	allocator.Delete()
-	allocator2.Delete()
-}
-
-func (s *AllocatorSuite) TestAllocateCached(c *C) {
-	testAllocator(c, idpool.ID(32), randomTestName(), "a") // enable use of local cache
-}
-
-func (s *AllocatorSuite) TestKeyToID(c *C) {
-	allocatorName := randomTestName()
-	a, err := NewAllocator(allocatorName, TestType(""), WithSuffix("a"))
-	c.Assert(err, IsNil)
-	c.Assert(a, Not(IsNil))
-
-	// An error is returned because the path is outside the prefix (allocatorName/id)
-	id, err := backend.keyToID(path.Join(allocatorName, "invalid"))
-	c.Assert(err, Not(IsNil))
-	c.Assert(id, Equals, idpool.NoID)
-
-	// An error is returned because the path contains the prefix
-	// (allocatorName/id) but cannot be parsed ("invalid")
-	id, err = backend.keyToID(path.Join(allocatorName, "id", "invalid"))
-	c.Assert(err, Not(IsNil))
-	c.Assert(id, Equals, idpool.NoID)
-
-	// A valid lookup that finds an ID
-	id, err = backend.keyToID(path.Join(allocatorName, "id", "10"))
-	c.Assert(err, IsNil)
-	c.Assert(id, Equals, idpool.ID(10))
-}
-
-func testGetNoCache(c *C, maxID idpool.ID, testName string, suffix string) {
-	allocator, err := NewAllocator(testName, TestType(""), WithMax(maxID),
-		WithSuffix(suffix), WithoutGC())
-	c.Assert(err, IsNil)
-	c.Assert(allocator, Not(IsNil))
-
-	// remove any keys which might be leftover
-	allocator.DeleteAllKeys()
-	defer allocator.DeleteAllKeys()
-
-	labelsLong := "foo;/;bar;"
-	key := TestType(fmt.Sprintf("%s%010d", labelsLong, 0))
-	longID, new, err := allocator.Allocate(context.Background(), key)
-	c.Assert(err, IsNil)
-	c.Assert(longID, Not(Equals), 0)
-	c.Assert(new, Equals, true)
-
-	observedID, err := allocator.GetNoCache(context.Background(), key)
-	c.Assert(err, IsNil)
-	c.Assert(observedID, Not(Equals), 0)
-
-	labelsShort := "foo;/;"
-	shortKey := TestType(labelsShort)
-	observedID, err = allocator.GetNoCache(context.Background(), shortKey)
-	c.Assert(err, IsNil)
-	c.Assert(observedID, Equals, idpool.NoID)
-
-	shortID, new, err := allocator.Allocate(context.Background(), shortKey)
-	c.Assert(err, IsNil)
-	c.Assert(shortID, Not(Equals), 0)
-	c.Assert(new, Equals, true)
-
-	observedID, err = allocator.GetNoCache(context.Background(), shortKey)
-	c.Assert(err, IsNil)
-	c.Assert(observedID, Equals, shortID)
-}
-
-func (s *AllocatorSuite) TestprefixMatchesKey(c *C) {
-	// cilium/state/identities/v1/value/label;foo;bar;/172.0.124.60
-
-	tests := []struct {
-		prefix   string
-		key      string
-		expected bool
-	}{
-		{
-			prefix:   "foo",
-			key:      "foo/bar",
-			expected: true,
-		},
-		{
-			prefix:   "foo/;bar;baz;/;a;",
-			key:      "foo/;bar;baz;/;a;/alice",
-			expected: true,
-		},
-		{
-			prefix:   "foo/;bar;baz;",
-			key:      "foo/;bar;baz;/;a;/alice",
-			expected: false,
-		},
-		{
-			prefix:   "foo/;bar;baz;/;a;/baz",
-			key:      "foo/;bar;baz;/;a;/alice",
-			expected: false,
-		},
-	}
-
-	for _, tt := range tests {
-		c.Logf("prefixMatchesKey(%q, %q) expected to be %t", tt.prefix, tt.key, tt.expected)
-		result := prefixMatchesKey(tt.prefix, tt.key)
-		c.Assert(result, Equals, tt.expected)
-	}
-}
-
-func (s *AllocatorSuite) TestGetNoCache(c *C) {
-	testGetNoCache(c, idpool.ID(256), randomTestName(), "a") // enable use of local cache
-}
-
-func (s *AllocatorSuite) TestRemoteCache(c *C) {
-	testName := randomTestName()
-	allocator, err := NewAllocator(testName, TestType(""), WithMax(idpool.ID(256)), WithSuffix("a"))
-	c.Assert(err, IsNil)
-	c.Assert(allocator, Not(IsNil))
-
-	// remove any keys which might be leftover
-	allocator.DeleteAllKeys()
-
-	// allocate all available IDs
-	for i := idpool.ID(1); i <= idpool.ID(4); i++ {
-		key := TestType(fmt.Sprintf("key%04d", i))
-		_, _, err := allocator.Allocate(context.Background(), key)
-		c.Assert(err, IsNil)
-	}
-
-	// wait for main cache to be populated
-	c.Assert(testutils.WaitUntil(func() bool { return len(allocator.mainCache.cache) == 4 }, 5*time.Second), IsNil)
-
-	// count identical allocations returned
-	cache := map[idpool.ID]int{}
-	allocator.ForeachCache(func(id idpool.ID, val AllocatorKey) {
-		cache[id]++
-	})
-
-	// ForeachCache must have returned 4 allocations all unique
-	c.Assert(len(cache), Equals, 4)
-	for i := range cache {
-		c.Assert(cache[i], Equals, 1)
-	}
-
-	// watch the prefix in the same kvstore via a 2nd watcher
-	rc := allocator.WatchRemoteKVStore(kvstore.Client(), testName)
-	c.Assert(rc, Not(IsNil))
-
-	// wait for remote cache to be populated
-	c.Assert(testutils.WaitUntil(func() bool { return len(rc.cache.cache) == 4 }, 5*time.Second), IsNil)
-
-	// count the allocations in the main cache *AND* the remote cache
-	cache = map[idpool.ID]int{}
-	allocator.ForeachCache(func(id idpool.ID, val AllocatorKey) {
-		cache[id]++
-	})
-
-	// Foreach must have returned 4 allocations each duplicated, once in
-	// the main cache, once in the remote cache
-	c.Assert(len(cache), Equals, 4)
-	for i := range cache {
-		c.Assert(cache[i], Equals, 2)
-	}
-
-	rc.Close()
-
-	allocator.DeleteAllKeys()
-	allocator.Delete()
-}
-
-// The following tests are currently disabled as they are not 100% reliable in
-// the Jenkins CI
 //
-//func testParallelAllocator(c *C, maxID idpool.ID, allocatorName string, suffix string) {
-//	allocator, err := NewAllocator(allocatorName, TestType(""), WithMax(maxID), WithSuffix(suffix))
+//type AllocatorEtcdSuite struct {
+//	AllocatorSuite
+//}
+//
+//var _ = Suite(&AllocatorEtcdSuite{})
+//
+//func (e *AllocatorEtcdSuite) SetUpTest(c *C) {
+//	kvstore.SetupDummy("etcd")
+//}
+//
+//func (e *AllocatorEtcdSuite) TearDownTest(c *C) {
+//	kvstore.DeletePrefix(testPrefix)
+//	kvstore.Close()
+//}
+//
+//type AllocatorConsulSuite struct {
+//	AllocatorSuite
+//}
+//
+//var _ = Suite(&AllocatorConsulSuite{})
+//
+//func (e *AllocatorConsulSuite) SetUpTest(c *C) {
+//	kvstore.SetupDummy("consul")
+//}
+//
+//func (e *AllocatorConsulSuite) TearDownTest(c *C) {
+//	kvstore.DeletePrefix(testPrefix)
+//	kvstore.Close()
+//}
+//
+////FIXME: this should be named better, it implements pkg/allocator.Backend
+//type TestType string
+//
+//func (t TestType) GetKey() string              { return string(t) }
+//func (t TestType) GetAsMap() map[string]string { return map[string]string{string(t): string(t)} }
+//func (t TestType) String() string              { return string(t) }
+//func (t TestType) PutKey(v string) (allocator.AllocatorKey, error) {
+//	return TestType(v), nil
+//}
+//func (t TestType) PutKeyFromMap(m map[string]string) allocator.AllocatorKey {
+//	for _, v := range m {
+//		return TestType(v)
+//	}
+//
+//	panic("empty map")
+//}
+//
+//func randomTestName() string {
+//	return testutils.RandomRuneWithPrefix(testPrefix, 12)
+//}
+//
+//func (s *AllocatorSuite) BenchmarkAllocate(c *C) {
+//	allocatorName := randomTestName()
+//	maxID := idpool.ID(256 + c.N)
+//	backend, err := NewKVStoreBackend(allocatorName, "a", TestType(""))
+//	c.Assert(err, IsNil)
+//	a, err := allocator.NewAllocator(TestType(""), backend, allocator.WithMax(maxID))
+//	c.Assert(err, IsNil)
+//	c.Assert(a, Not(IsNil))
+//	defer a.DeleteAllKeys()
+//
+//	c.ResetTimer()
+//	for i := 0; i < c.N; i++ {
+//		_, _, err := a.Allocate(context.Background(), TestType(fmt.Sprintf("key%04d", i)))
+//		c.Assert(err, IsNil)
+//	}
+//	c.StopTimer()
+//
+//}
+//
+//func (s *AllocatorSuite) TestGC(c *C) {
+//	allocatorName := randomTestName()
+//	maxID := idpool.ID(256 + c.N)
+//	// FIXME: Did this previousy use allocatorName := randomTestName() ? so TestType(randomeTestName())
+//	backend, err := NewKVStoreBackend(allocatorName, "a", TestType(""))
+//	c.Assert(err, IsNil)
+//	allocator, err := allocator.NewAllocator(TestType(""), backend, allocator.WithMax(maxID), allocator.WithoutGC())
 //	c.Assert(err, IsNil)
 //	c.Assert(allocator, Not(IsNil))
+//	defer allocator.DeleteAllKeys()
+//	defer allocator.Delete()
+//
+//	allocator.DeleteAllKeys()
+//
+//	shortKey := TestType("1;")
+//	shortID, _, err := allocator.Allocate(context.Background(), shortKey)
+//	c.Assert(err, IsNil)
+//	c.Assert(shortID, Not(Equals), 0)
+//
+//	longKey := TestType("1;2;")
+//	longID, _, err := allocator.Allocate(context.Background(), longKey)
+//	c.Assert(err, IsNil)
+//	c.Assert(longID, Not(Equals), 0)
+//
+//	allocator.Release(context.Background(), shortKey)
+//
+//	keysToDelete := map[string]uint64{}
+//	keysToDelete, err = allocator.RunGC(keysToDelete)
+//	c.Assert(err, IsNil)
+//	c.Assert(len(keysToDelete), Equals, 1)
+//	keysToDelete, err = allocator.RunGC(keysToDelete)
+//	c.Assert(err, IsNil)
+//	c.Assert(len(keysToDelete), Equals, 0)
+//
+//	// wait for cache to be updated via delete notification
+//	c.Assert(testutils.WaitUntil(func() bool {
+//		key, err := allocator.GetByID(shortID)
+//		if err != nil {
+//			c.Error(err)
+//			return false
+//		}
+//		//FIXME: This isn't nil but it probably should be. This is because TestType is setup with "" and returns itself, a non-nil
+//		return key == nil || key.String() == ""
+//	}, 5*time.Second), IsNil)
+//
+//	key, err := allocator.GetByID(shortID)
+//	c.Assert(err, IsNil)
+//	c.Assert(key, Equals, TestType(""))
+//}
+//
+//func testAllocator(c *C, maxID idpool.ID, allocatorName string, suffix string) {
+//	backend, err := NewKVStoreBackend(allocatorName, "a", TestType(""))
+//	c.Assert(err, IsNil)
+//	a, err := allocator.NewAllocator(TestType(""), backend,
+//		allocator.WithMax(maxID), allocator.WithoutGC())
+//	c.Assert(err, IsNil)
+//	c.Assert(a, Not(IsNil))
+//
+//	// remove any keys which might be leftover
+//	a.DeleteAllKeys()
 //
 //	// allocate all available IDs
 //	for i := idpool.ID(1); i <= maxID; i++ {
 //		key := TestType(fmt.Sprintf("key%04d", i))
-//		id, _, err := allocator.Allocate(context.Background(), key)
+//		id, new, err := a.Allocate(context.Background(), key)
 //		c.Assert(err, IsNil)
 //		c.Assert(id, Not(Equals), 0)
-//
-//		// refcnt must be 1
-//		c.Assert(allocator.localKeys.keys[key.GetKey()].refcnt, Equals, uint64(1))
+//		c.Assert(new, Equals, true)
 //	}
-//
-//	saved := allocator.backoffTemplate.Factor
-//	allocator.backoffTemplate.Factor = 1.0
-//
-//	// we should be out of id space here
-//	_, new, err := allocator.Allocate(context.Background(), TestType(fmt.Sprintf("key%04d", maxID+1)))
-//	c.Assert(err, Not(IsNil))
-//	c.Assert(new, Equals, false)
-//
-//	allocator.backoffTemplate.Factor = saved
 //
 //	// allocate all IDs again using the same set of keys, refcnt should go to 2
 //	for i := idpool.ID(1); i <= maxID; i++ {
 //		key := TestType(fmt.Sprintf("key%04d", i))
-//		id, _, err := allocator.Allocate(context.Background(), key)
+//		id, new, err := a.Allocate(context.Background(), key)
 //		c.Assert(err, IsNil)
 //		c.Assert(id, Not(Equals), 0)
-//
-//		// refcnt must now be 2
-//		c.Assert(allocator.localKeys.keys[key.GetKey()].refcnt, Equals, uint64(2))
+//		c.Assert(new, Equals, false)
 //	}
 //
+//	// Create a 2nd allocator, refill it
+//	backend2, err := NewKVStoreBackend(allocatorName, "r", TestType(""))
+//	c.Assert(err, IsNil)
+//	a2, err := allocator.NewAllocator(TestType(""), backend2,
+//		allocator.WithMax(maxID), allocator.WithoutGC())
+//	c.Assert(err, IsNil)
+//	c.Assert(a2, Not(IsNil))
+//
+//	// allocate all IDs again using the same set of keys, refcnt should go to 2
 //	for i := idpool.ID(1); i <= maxID; i++ {
-//		allocator.Release(context.Background(), TestType(fmt.Sprintf("key%04d", i)))
+//		key := TestType(fmt.Sprintf("key%04d", i))
+//		id, new, err := a2.Allocate(context.Background(), key)
+//		c.Assert(err, IsNil)
+//		c.Assert(id, Not(Equals), 0)
+//		c.Assert(new, Equals, false)
+//
+//		a2.Release(context.Background(), key)
 //	}
+//
+//	// release 2nd reference of all IDs
+//	for i := idpool.ID(1); i <= maxID; i++ {
+//		a.Release(context.Background(), TestType(fmt.Sprintf("key%04d", i)))
+//	}
+//
+//	staleKeysPreviousRound := map[string]uint64{}
+//	// running the GC should not evict any entries
+//	staleKeysPreviousRound, err = a.RunGC(staleKeysPreviousRound)
+//	c.Assert(err, IsNil)
+//
+//	v, err := kvstore.ListPrefix(path.Join(allocatorName, "id"))
+//	c.Assert(err, IsNil)
+//	c.Assert(len(v), Equals, int(maxID))
 //
 //	// release final reference of all IDs
 //	for i := idpool.ID(1); i <= maxID; i++ {
-//		allocator.Release(context.Background(), TestType(fmt.Sprintf("key%04d", i)))
+//		a.Release(context.Background(), TestType(fmt.Sprintf("key%04d", i)))
 //	}
 //
 //	// running the GC should evict all entries
-//	allocator.RunGC()
+//	staleKeysPreviousRound, err = a.RunGC(staleKeysPreviousRound)
+//	c.Assert(err, IsNil)
+//	_, err = a.RunGC(staleKeysPreviousRound)
+//	c.Assert(err, IsNil)
 //
-//	v, err := kvstore.ListPrefix(allocator.idPrefix)
+//	v, err = kvstore.ListPrefix(path.Join(allocatorName, "id"))
 //	c.Assert(err, IsNil)
 //	c.Assert(len(v), Equals, 0)
 //
-//	allocator.Delete()
+//	a.DeleteAllKeys()
+//	a.Delete()
+//	a2.Delete()
 //}
 //
-//func (s *AllocatorSuite) TestParallelAllocation(c *C) {
-//	var (
-//		wg            sync.WaitGroup
-//		allocatorName = randomTestName()
-//	)
+//func (s *AllocatorSuite) TestAllocateCached(c *C) {
+//	testAllocator(c, idpool.ID(32), randomTestName(), "a") // enable use of local cache
+//}
 //
-//	// create dummy allocator to delete all keys
-//	a, err := NewAllocator(allocatorName, TestType(""), WithSuffix("a"))
+//func (s *AllocatorSuite) TestKeyToID(c *C) {
+//	allocatorName := randomTestName()
+//	backend, err := NewKVStoreBackend(allocatorName, "a", TestType(""))
+//	c.Assert(err, IsNil)
+//	a, err := allocator.NewAllocator(TestType(""), backend)
 //	c.Assert(err, IsNil)
 //	c.Assert(a, Not(IsNil))
-//	defer a.DeleteAllKeys()
-//	defer a.Delete()
 //
-//	for i := 0; i < 2; i++ {
-//		wg.Add(1)
-//		go func() {
-//			defer wg.Done()
-//			testParallelAllocator(c, idpool.ID(64), allocatorName, fmt.Sprintf("node-%d", i))
-//		}()
+//	// An error is returned because the path is outside the prefix (allocatorName/id)
+//	id, err := backend.keyToID(path.Join(allocatorName, "invalid"))
+//	c.Assert(err, Not(IsNil))
+//	c.Assert(id, Equals, idpool.NoID)
+//
+//	// An error is returned because the path contains the prefix
+//	// (allocatorName/id) but cannot be parsed ("invalid")
+//	id, err = backend.keyToID(path.Join(allocatorName, "id", "invalid"))
+//	c.Assert(err, Not(IsNil))
+//	c.Assert(id, Equals, idpool.NoID)
+//
+//	// A valid lookup that finds an ID
+//	id, err = backend.keyToID(path.Join(allocatorName, "id", "10"))
+//	c.Assert(err, IsNil)
+//	c.Assert(id, Equals, idpool.ID(10))
+//}
+//
+//func testGetNoCache(c *C, maxID idpool.ID, suffix string) {
+//	allocatorName := randomTestName()
+//	backend, err := NewKVStoreBackend(allocatorName, "a", TestType(""))
+//	c.Assert(err, IsNil)
+//	allocator, err := allocator.NewAllocator(TestType(""), backend, allocator.WithMax(maxID), allocator.WithoutGC())
+//	c.Assert(err, IsNil)
+//	c.Assert(allocator, Not(IsNil))
+//
+//	// remove any keys which might be leftover
+//	allocator.DeleteAllKeys()
+//	defer allocator.DeleteAllKeys()
+//
+//	labelsLong := "foo;/;bar;"
+//	key := TestType(fmt.Sprintf("%s%010d", labelsLong, 0))
+//	longID, new, err := allocator.Allocate(context.Background(), key)
+//	c.Assert(err, IsNil)
+//	c.Assert(longID, Not(Equals), 0)
+//	c.Assert(new, Equals, true)
+//
+//	observedID, err := allocator.GetNoCache(context.Background(), key)
+//	c.Assert(err, IsNil)
+//	c.Assert(observedID, Not(Equals), 0)
+//
+//	labelsShort := "foo;/;"
+//	shortKey := TestType(labelsShort)
+//	observedID, err = allocator.GetNoCache(context.Background(), shortKey)
+//	c.Assert(err, IsNil)
+//	c.Assert(observedID, Equals, idpool.NoID)
+//
+//	shortID, new, err := allocator.Allocate(context.Background(), shortKey)
+//	c.Assert(err, IsNil)
+//	c.Assert(shortID, Not(Equals), 0)
+//	c.Assert(new, Equals, true)
+//
+//	observedID, err = allocator.GetNoCache(context.Background(), shortKey)
+//	c.Assert(err, IsNil)
+//	c.Assert(observedID, Equals, shortID)
+//}
+//
+//func (s *AllocatorSuite) TestprefixMatchesKey(c *C) {
+//	// cilium/state/identities/v1/value/label;foo;bar;/172.0.124.60
+//
+//	tests := []struct {
+//		prefix   string
+//		key      string
+//		expected bool
+//	}{
+//		{
+//			prefix:   "foo",
+//			key:      "foo/bar",
+//			expected: true,
+//		},
+//		{
+//			prefix:   "foo/;bar;baz;/;a;",
+//			key:      "foo/;bar;baz;/;a;/alice",
+//			expected: true,
+//		},
+//		{
+//			prefix:   "foo/;bar;baz;",
+//			key:      "foo/;bar;baz;/;a;/alice",
+//			expected: false,
+//		},
+//		{
+//			prefix:   "foo/;bar;baz;/;a;/baz",
+//			key:      "foo/;bar;baz;/;a;/alice",
+//			expected: false,
+//		},
 //	}
 //
-//	wg.Wait()
+//	for _, tt := range tests {
+//		c.Logf("prefixMatchesKey(%q, %q) expected to be %t", tt.prefix, tt.key, tt.expected)
+//		result := prefixMatchesKey(tt.prefix, tt.key)
+//		c.Assert(result, Equals, tt.expected)
+//	}
+//}
+//
+//func (s *AllocatorSuite) TestGetNoCache(c *C) {
+//	testGetNoCache(c, idpool.ID(256), "a") // enable use of local cache
+//}
+//
+//func (s *AllocatorSuite) TestRemoteCache(c *C) {
+//	testName := randomTestName()
+//	backend, err := NewKVStoreBackend(testName, "a", TestType(""))
+//	c.Assert(err, IsNil)
+//	a, err := allocator.NewAllocator(TestType(""), backend, allocator.WithMax(idpool.ID(256)))
+//	c.Assert(err, IsNil)
+//	c.Assert(a, Not(IsNil))
+//
+//	// remove any keys which might be leftover
+//	a.DeleteAllKeys()
+//
+//	// allocate all available IDs
+//	for i := idpool.ID(1); i <= idpool.ID(4); i++ {
+//		key := TestType(fmt.Sprintf("key%04d", i))
+//		_, _, err := a.Allocate(context.Background(), key)
+//		c.Assert(err, IsNil)
+//	}
+//
+//	// wait for main cache to be populated
+//	c.Assert(testutils.WaitUntil(func() bool {
+//		cacheLen := 0
+//		a.ForeachCache(func(id idpool.ID, val allocator.AllocatorKey) {
+//			cacheLen++
+//		})
+//		return cacheLen == 4
+//	}, 5*time.Second), IsNil)
+//
+//	// count identical allocations returned
+//	cache := map[idpool.ID]int{}
+//	a.ForeachCache(func(id idpool.ID, val allocator.AllocatorKey) {
+//		cache[id]++
+//	})
+//
+//	// ForeachCache must have returned 4 allocations all unique
+//	c.Assert(len(cache), Equals, 4)
+//	for i := range cache {
+//		c.Assert(cache[i], Equals, 1)
+//	}
+//
+//	// watch the prefix in the same kvstore via a 2nd watcher
+//	rc := a.WatchRemoteKVStore(kvstore.Client(), testName)
+//	c.Assert(rc, Not(IsNil))
+//
+//	// wait for remote cache to be populated
+//	c.Assert(testutils.WaitUntil(func() bool {
+//		cacheLen := 0
+//		a.ForeachCache(func(id idpool.ID, val allocator.AllocatorKey) {
+//			cacheLen++
+//		})
+//		// 4 local + 4 remote
+//		return cacheLen == 8
+//	}, 5*time.Second), IsNil)
+//
+//	// count the allocations in the main cache *AND* the remote cache
+//	cache = map[idpool.ID]int{}
+//	a.ForeachCache(func(id idpool.ID, val allocator.AllocatorKey) {
+//		cache[id]++
+//	})
+//
+//	// Foreach must have returned 4 allocations each duplicated, once in
+//	// the main cache, once in the remote cache
+//	c.Assert(len(cache), Equals, 4)
+//	for i := range cache {
+//		c.Assert(cache[i], Equals, 2)
+//	}
+//
+//	rc.Close()
+//
+//	a.DeleteAllKeys()
+//	a.Delete()
 //}
